chore(components): update all input imports
