+++ src/app/auth/sign-up/_components/sign-up-form.tsx
@@ -4,7 +4,7 @@ import { Form } from '@/components/form'
 import { Alert } from '@/components/ui/alert'
 import { Button } from '@/components/ui/button'
 import { Checkbox } from '@/components/ui/checkbox-new'
-import { Input } from '@/components/ui/input-new'
+import { Input } from '@/components/ui/input'
 import { ShowAndHideButton } from '@/components/ui/show-and-hide-button'
 
 import { MIN_PASSWORD_LENGTH } from '../_helpers/sign-up-form-validation'
